feat:separate out component from the main page and add share page
