+++ package.json
@@ -17,6 +17,7 @@
     "@radix-ui/react-scroll-area": "^1.2.0",
     "@radix-ui/react-slot": "^1.1.0",
     "@radix-ui/react-tabs": "^1.1.0",
+    "@radix-ui/react-toast": "^1.2.2",
     "@redheadphone/react-json-grid": "^0.7.0",
     "@tanstack/react-query": "^5.59.0",
     "class-variance-authority": "^0.7.0",
