+++ src/app/s/[id]/page.tsx
@@ -1,0 +1,11 @@
+import { JsonVisualizer } from "@/components/json-visualizer";
+
+interface SharedPageProps {
+  params: {
+    id: string;
+  };
+}
+
+export default function SharedJsonPage({ params }: SharedPageProps) {
+  return <JsonVisualizer initialShareId={params.id} />;
+}
